scripts/requirements.txt
cachetools==2.1.0
certifi==2018.8.13
chardet==3.0.4
google-api-python-client==1.7.4
google-auth==1.5.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.2.0
httplib2==0.11.3
idna==2.7
oauthlib==2.1.0
pkg-resources==0.0.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
requests==2.19.1
requests-oauthlib==1.0.0
rsa==3.4.2
six==1.11.0
uritemplate==3.0.0
urllib3==1.23
